refactor(dive-log): extract and parameterize DiveFilterSheet
Move DiveFilterSheet + _DiveFilterSheetState out of dive_list_page.dart
into their own file so the sheet can target either the dive-list filter
provider (default) or another StateProvider<DiveFilterState>, e.g. the
upcoming Statistics filter.

Deviations from the task brief, both required for correctness:

- The brief's proposed `this.filterProvider = diveFilterProvider` default
  parameter value does not compile: diveFilterProvider is a `final`
  top-level variable, not `const`, and Dart requires default parameter
  values to be compile-time constants. Fixed by making the constructor
  parameter nullable (defaulting to null, a valid constant) and resolving
  it to diveFilterProvider via the constructor initializer list, which
  keeps the field itself non-nullable so `widget.filterProvider` is used
  unchanged everywhere else. The constructor can no longer be `const`,
  but no call site ever invoked it as const (they all pass a runtime
  `ref`), so this is not an observable behavior change.
- Also fixed the pre-existing coverage test
  test/features/dive_log/presentation/pages/dive_filter_sheet_test.dart,
  which imported DiveFilterSheet from dive_list_page.dart; that import
  now points at the new dive_filter_sheet.dart file.

The three existing call sites (dive_list_page.dart, dive_list_content.dart
x2) are unchanged (`DiveFilterSheet(ref: ref)`); dive list behavior is
unaffected.
